Weekly Chart Works
Its white and currently is displaying 2 weeks ahead (for testing purposes). Need to create monthly graph. I think everything can just be in HomeFragment because it looks like its simpler like that. The setup and customization should be the same and the only difference should be the list that is given by the monthly/weekly. Almost done with HomeFragment.