home page is done except for the biggest income category
I have idea on how to solve it. Just simplify it. Instead of calling IncomeCategoryData, call IncomeData. We are overcomplicating it right now.

Need to fix the item borders, need to figure out whether thick grey and thin accent or thin grey and thick accent.

Need to add white container cards to all recycler views.

Add character limits to the names of the income categories and expense categories.

Add Fonts